remove menu underline animation
this is because the animation can't follow the scroll page behaviour i.e. from /#experience (middle of page) to /not-found (top of page), it will move from left to right fine but will appear moving from bottom to top as well i.e. not aligned with the nav bar/menu/menu items